after_sucess

+++ Rakefile
@@ -21,21 +21,23 @@ namespace :push do
   desc 'push built site to s3 branch'
   task :branch do
     if ENV['CI']
-      REPO_SLUG = ENV['TRAVIS_REPO_SLUG']
       BRANCH = ENV['TRAVIS_BRANCH']
-      USER = REPO_SLUG.split('/')[0]
-      TOKEN = ENV['ACCESS_TOKEN']
-      COMMIT_MSG = "Site updated via #{ENV['TRAVIS_COMMIT']}".freeze
-      ORIGIN = "https://#{USER}:#{TOKEN}@github.com/#{REPO_SLUG}.git".freeze
-      puts "Deploying to #{BRANCH}_site from Travis as #{USER}"
+      if BRANCH == 'production' or 'staging'
+        REPO_SLUG = ENV['TRAVIS_REPO_SLUG']
+        USER = REPO_SLUG.split('/')[0]
+        TOKEN = ENV['ACCESS_TOKEN']
+        COMMIT_MSG = "Site updated via #{ENV['TRAVIS_COMMIT']}".freeze
+        ORIGIN = "https://#{USER}:#{TOKEN}@github.com/#{REPO_SLUG}.git".freeze
+        puts "Deploying to #{BRANCH}_site from Travis as #{USER}"
 
-      Dir.mktmpdir do |tmp|
-        cp_r '_site/.', tmp
-        Dir.chdir tmp
-        system 'git init'
-        system "git add . && git commit -m '#{COMMIT_MSG}'"
-        system "git remote add origin #{ORIGIN}"
-        system "git push origin master:refs/heads/#{BRANCH}_site --force"
+        Dir.mktmpdir do |tmp|
+          cp_r '_site/.', tmp
+          Dir.chdir tmp
+          system 'git init'
+          system "git add . && git commit -m '#{COMMIT_MSG}'"
+          system "git remote add origin #{ORIGIN}"
+          system "git push origin master:refs/heads/#{BRANCH}_site --force"
+        end
       end
     end
   end
